Added null objects and created transaction models
Added null objects for users, items, accounts and transactions
Created a transaction model and added table to database

Note:
In the commands file, I previously had a transactions folder with
a 'Transactions::FindNewTransactions' module. This was causing conflicts when running
'rails g model transaction'. The command line told me "Transaction" was
already being used in my rails application or a reserved word by Ruby on
Rails. Changing the training folder to 'place_holder' and the module to
'PlaceHolder::FindNewTransactions' allowed me to create the model